misc/reboot: move to cmd/internal/bootstrap_test
This is the last test run from misc by dist.

For #37486.

Reviewed-on: https://go-review.googlesource.com/c/go/+/493606
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Austin Clements <austin@google.com>
